requirements.txt
Twisted==14.0.0
autobahn==0.8.9
cffi==0.8.2
cryptography==0.4
mock==1.0.1
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
service-identity==0.2
six==1.6.1
zope.interface==4.1.1
